Refactor getCompany & getCompanies to it's own business folder for easier testing
